indexcodec: Implement next restart in version=1 stream
When we detect 0/1/2 triangle we emit a special restart code, that
resets next back to 0 so that we can continue the encoding sequence.

To encode this, we use the redundancy where codeaux 00 (which logically
represents the triangle we're resetting this anyhow) is usually encoded
as a table entry; we thus encode it as a non-table entry which indicates
that in addition to the usual decoding we need to reset next.