Add array deduction guides (#343)
* Add array deduction guides

Allows inference of an array's type and length from the types and
number of elements in its constructor. The guides are defined as
part of the C++17 standard, and are gated behind a check for
the __cpp_deduction_guides define as per the span implementation.

* Change value_type test to static_assert

Wasn't initially sure if static_asserts were okay in tests or if
they should all go through the VERIFY() macro at runtime, however
upon looking around this seems to be fine.