indexcodec: Add a way to set encoding version
To improve compression ratio, we need to change the bitstream which
requires introducing a new bitstream version.

While decoding will automatically work with any encoded version, when
encoding we need to specify whether we want a backwards-compatible
result (default) or the one that results in better compression.

For now this is managed using a global variable; we'll see if this is
the best way going forward.