indexcodec: Improve strip encoding further
In the previous change, the index bitstream format was extended to
handle last-1 vertex references within a single triangle byte.

This helps for strips that go backwards, but doesn't help for strips
that go forwards! Because of this the encoding of large planar patches
was actually suboptimal.

gzip cleans up most of this because even without the special encoding
the emitted bytes are very repetitive, but we'd like to get some savings
even before gzip.

This change steals one more value from the FIFO back-reference to encode
last+1. This results in 1-10% of savings pre-deflate for strip-encoded
index data without a noticeable difference in performance.

Additionally this fixes a bug introduced in Makefile where the _base
WASM variant wasn't being patched into .js file properly.

Note that this *changes* the encoding with the version 1 - however,
version 1 encoding is not released yet so it should be fine to iterate
on it in master.