C++20 char8_t string/string_view support improvements (#345)
* Add char8_t support for string_view's ""sv operators

Enables the ""sv and ""_sv operators if the compiler implements
support for the distinct char8_t type in present in C++20.

Tests are included which *should* trigger under the MSVC part of
the CI setup; I've verified locally that it all works with Clang 9
and GCC 9.2.0 with a local hack to compile only the string_view
tests in C++20 mode.

Includes a specialization of hash<u8string_view> to enable the
tests to actually compile successfully.

* Improve char8_t support for strings

When char8_t support is enabled within a compiler, include a
literal ""s operator for converting a u8"" string literal to
an instance of u8string.

Enabled tests for u8string if char8_t is a distinct type, which
as with the string_view support required another hash specialization
to be defined, as well as a few extra overloads of DecodePart to
prevent compilation errors in tests that append strings of mixed
encodings. Each added overload simply casts the char8_t* data to
char* and defers to another overload to handle the operation.

* Enable commented-out wchar_t string tests

These can be supported with an additional DecodePart overload
as per the previous char8_t changeset to support writing data
from and to a buffer of the same type. This is implemented just
as a straight cast to the char* overload which does a memmove,
so should be safe.

* Add EASTL_BUILD_CHAR8T_TESTS option to test build

This enables building a subset of the tests (currently string
and string_view) with char8_t support if the compiler being
used supports either the `-fchar8_t` or `/Zc:char8_t` flag.

If this option is enabled a diagostic will be printed if the
flag is successfully found or not.

* Add Linux Clang build to matrix for char8_t support

The Clang compiler used in the environment (Clang 7) supposedly
supports the `-fchar8_t` flag so this should be a good way to
prove that things work.

The MSVC compiler under the Travis environment *doesn't* support
/Zc:char8_t at this time; rather than remove that from the matrix
it's been left there since removing it is a bit out-of-scope of
this changeset.

* Fail test compilation if char8_t support is missing

If the tests were compiled with char8_t support requested, and the
appropriate compiler flag was found, we set a define which can be
tested for to trigger a compile error if EA_CHAR8_UNIQUE still
evaluates to false.

This should work to prove that the tests are correctly compiling
with char8_t support enabled under the CI setup when the CMake
configure process says it's doing so.

* Remove clang char8_t build from matrix

* Unconditionally enable char8_t tests if supported

The string/string_view tests will be built with char8_t support
if the compiler flag enabling it is present. The test for the flag
was moved to CommonCppFlags since we'll need it in the benchmarks
file to fix a compilation failure there under recent MSVC versions.

* Build benchmarks without char8_t support

Under recent versions of MSVC, the use of the `/permissive-` flag
in CommonCppFlags enables char8_t support which caused compilation
failures with the tests and benchmarks.

As this isn't supported yet, if the compiler supports a flag that
explicitly disables char8_t support (`-fno-char8_t` or `/Zc:char8_t-`)
then we compile both the tests and benchmarks with that flag by
default, overriding only for specific files that do support it.

Use of add_compile_options instead of `set(CMAKE_CXX_FLAGS ...)` is
to ensure the flag propagates down to the submodules, since for some
reason this wasn't happening reliably.