gltfpack: Implement support for variable # of normal bits
This change reworks the approach to encoding normal/tangent vectors.

First, we no longer have "unnormalized" setting. When normal filters are
used, normals are always normalized, otherwise they're normalized if
glTF allows for it (8/16 bits) and not normalized otherwise.

Second, we now limit the tangent vector to 8 bits - there's not much
sense in going >8 bits for tangents, but more than 8 bits might be
important for normals. This might change in the future, for now this
simplifies the code a bit.

Third, we now use the specified normal bit count for encoding with or
without filters. This results in more intuitive behavior, and allows
users to trade off quality for size. The default is still 8 as it should
be appropriate for most models.