vertexcodec: Fix empty buffer encoding
Previously when vertex count was 0, we still read one vertex worth of
data from the input data. This could result in a crash due to reading
out of bounds memory.

This change fixes it and adds a unit test for this case - the case is
admittedly very rare but gltfpack hits it on one of the unusual test
files from glTF-Validator repository.