gltfpack: Write material extras to the output file
This change introduces -ke flag that is used to keep the extras attached
to materials (and, in the future, other data).

This is conditional because without knowing anything about the extras
content we can't tell if it's going to be valid or not - e.g. it could
refer to some objects that we have removed or reordered. In the future
this flag might need to block some of the optimizations.