gltfpack: Restrict material merging when -ke is in effect
When merging equivalent materials, we compare extras data and only do so
when it's equivalent.

Note that ideally we would compare minified data or perform
structure-aware comparison, but a basic comparison is robust and
probably as accurate in practice given that we can't reason about the
data inside extras anyhow.