Update WASM kernels to be more ARM-friendly
Instead of using swizzle and half-shuffle (neither of which are
representable on ARM64), we encode splat directly and use unziplo/unziphi
which is representable as shufps on x64 and unzip on ARM.

Note that this doesn't change the JS codegen - it's the same shuffles,
just spelled out differently to prevent future misuse.

Also remove pblendw-like shuffle because that's slow on ARM as well
right now.