gltfpack: Implement support for shared exponent encoding
Instead of using quantized floats for translation/scale animation
tracks, we use a shared exponent format when -cc is specified.

To make the number of bits easier to tune, we now specify the number of
mantissa bits (including the sign), so the valid range is up to 24 bits.

When -cc is not specified, we now output raw values - this is consistent
with our general approach where -c uses maximum precision of the output
data, matching in-memory and transmission precision, whereas -cc trades
off additional precision to reduce transmission size.

This can make translation tracks ~10-20% smaller for most models that
animate translations compared to our old float quantization, with bigger
wins for translation tracks on typical character models.