js: Use more efficient .wasm embedding
Instead of simply encoding input .wasm data as hexadecimal strings, we
now encode the input as a 6-bit stream with Base64.

To keep the data compressible, we need to encode individual bytes to
keep LZ matches; to achieve this, we precompute a frequency table and
encode most frequent 60 elements as is, and the rest as two bytes (2 + 6
bits). Around 93% of inputs bytes in Wasm blobs can be encoded using 6
bits, which results in the encoded stream being close in size to the
source byte stream, and the result compresses well with .gzip.

As a result of this change, both .js and .js.gz files are smaller:

.js: before 35103, after 20714 (-41%)
.js.gz: before 6444, after 6069 (-6%)

This is including the new unpack function.