gltfpack: Free unused buffers after glTF parsing
After we parse mesh and animation data, we can usually free the buffer
data. The exceptions are when the buffers are referenced by images or
skins, since we don't copy these out.

We also can't free the GLB buffers yet because this requires freeing
JSON data which requires fixing extras[] access.

This reduces max RSS when processing Thai Buddha scene from 1198 MB to
893 MB.