gltfpack: Don't generate fallback buffer unless required
Instead of always generating the fallback buffer, we only do this when
the fallback is actually requested. This saves us some time to copy the
uncompressed data, and means we don't need to create uncompressed
fallback in memory just to throw it away.

This reduces max RSS when processing Thai Buddha from 893 MB to 822 MB.