gltfpack: Free .gltf / .glb data early
Since we now extract extras into a separate string, we can free .gltf
file contents immediately after parsing it. If we're parsing a .glb
file, we need to hold onto the buffer data until we parse it fully, but
once we do - if we don't have any skin/image data inside - we can free
it as well.

This reduces the max RSS for stadium.glb from 100 MB to 75 MB.