simplify: Use bidirectional edge loops
It turns out that edge loop data is exactly the same as open edge data.
loop[] is equivalent to openout[], and openinc[] - which we didn't have
before - is equivalent to loopback[], which points in the opposite
direction of the edge winding.

We don't need loopback[] during simplification right now, but remapping
loops is cheap; we can use the same remapping process by treating the
loopback[] as a unidirectional loop.