simplify: Improve edge quadrics to add errors for corner edges
When a border-border edge is at the end of an edge loop, the next vertex
along the loop isn't a border vertex - fillEdgeQuadrics previously
ignored these edges.

It's incorrect to do so, however, since this means that we aren't
recording the error information on some corners, e.g. for the case below
where A, D are locked and B, C are border, failure to record AB or CD
errors would result in the edge BC having very small collapse error in
either direction:

	--A  D--
	  |  |
	  B--C

This change fixes the problem by using loopback[] to verify the loop
connection. This fixes some cases where our collapses didn't preserve
appearance near complex topology, #122 being the motivating example.