gltfpack: Implement new way of texture resizing
Instead of relying on toktx to resize the texture we now read the
width/height of the source texture by parsing the PNG header and create
a scaled and rounded size from it.

For rounding, when -tp is active, we round to the nearest power of 2
instead of the next power of 2 to prevent excessive growth of texture
sizes.

When -tp is not active, we round to the next boundary of 4 as this is
the restriction imposed by many different APIs, notably Direct3D and, as
an extension, WebGL, when loading block-compressed content.