gltfpack: Fix output animation start time
Due to a bug in resampling logic, minimum time value was always 0 - this
could result in animations that were too long, potentially causing
issues for the client side playback.

Note that if there are no tracks, mint stays at FLT_MAX and we correct
it by clamping it to max (so both end up 0) - this shouldn't happen
normally, and animations without tracks shouldn't be output, but it can
happen in rare edge cases and it's better to fix this up early instead
of triggering issues with timing logic later.

Reported in #185.