gltfpack: Implement support for custom color bit count
8 bit colors are not always enough, especially since glTF uses linear
storage for vertex colors which is suboptimal - in general ~10 bits are
required for linear storage to eliminate visible banding for 8-bit sRGB
output, and more bits are required if the output is higher than 8 bits.

This change implements support for custom bit counts for color
components. To maximize the compression ratio, when bit count isn't 8 or
16, the bottom bits of the resulting value are set to 0 or 1 based on
the high bit. Varying the bit is important as this preserves 0 and 1
exactly.

A more correct scheme would involve replicating multiple top bits into
the bottom bits, but this results in significantly worse compression
ratio for N > 8, as the compressor can't exploit the corellation between
the high and low bytes. With 0 or 1 simply replicated, the byte
structure is easier to discern and the result compresses better.