vertexfilter: Fix decoding of data that's not aligned by 4
When count isn't divisible by 4, we need to process the tail separately
when using SIMD implementations. This was previously unimplemented,
which wasn't an issue for JS use because it was simple to pad the data
before decoding.

When using the filters in native code however, padding might prove
problematic, esp. when interfacing with cgltf as it allocates the data
for the caller.

Instead of dispatching to a scalar tail, which can result in larger code
size, we copy the tail to stack, run the processing on the final 4
elements and copy the results back.