tools: Add fuzz driver for all codecs
All codecs must be resistant to malicious inputs; while the decoders
aren't guaranteed to provide a valid result, they must be able to decode
the input without reading past the input bounds or writing past the
output bounds.

This change adds a fuzzing driver that attempts to prove this for both
SIMD and scalar implementations of the decoders.