simplify: Initial implementation of flip prevention
Right now simplifier does not restrict edge collapses beyond
topology/quadric errors, which in some cases results in triangles
getting flipped.

This is a well known problem and a well known solution is to check
whether the triangles are flipped explicitly during collapse.

Unfortunately, the simplifier currently doesn't maintain enough data to
be able to perform this analysis; this change adds code to maintain a
second adjacency structure (that identifies triangles during collapse)
and uses it to validate collapses right before they happen.

We only do this for collapses as they're about to be performed, which
can sometimes result in suboptimal results as we will only do the check
once per edge for bidirectional collapses, but as it stands this is very
expensive and more effort will need to be spent to minimize the
problems.

The current implementation seems to resolve existing issues with the
flipping but results in substantially worse simplification performance
(4-5x slowdown, most of the work seems to be in hasTriangleFlips).