simplify: Introduce an aggressive flip prevention mode
This code locks the ring of vertices around the collapse to avoid issues
with adjacency becoming stale and vertex movement not restricting the
flips in subsequent collapses.

To make debugging issues with simplification & face winding easier, we
can now output an .obj file with a freshly computed set of normals.