simplify: Return relative error after simplification
meshopt_simplify now has an extra parameter that can be used to return
the error after the simplification concludes. This can be useful to
estimate the distance at which to switch the levels of detail.

Instead of using the old approximate pass counting mechanism we just
track the accurate value throughout simplification process; as usual,
internally we work with quadratic errors and have to linearize it at the
end.