simplify: Add input & output error to simplifySloppy
This change makes simplifySloppy interface match that of simplify, so
that an error limit can be specified up front, and the resulting error
is returned when requested.

As a consequence, simplifySloppy can no longer guarantee that the output
will be at or under the requested triangle count; this unifies the
interface between simplify & simplifySloppy further and gives us some
opportunity to compute triangle count in a more approximate way in the
future.

As a result we may need to compute the triangle count for the min. grid
size before simplification starts; this sometimes increases the number
of passes we need to do by 1, but in aggregate actually *decreases* the
number of passes as the interpolation search gets more data and can be
more precise.