clusterizer: Refactor various parts of implementation
This change generally cleans up the new code to make it production
ready. The kD-tree code is now a bit more generic and can be used to
look up vertices or triangles if need be; vertex based code can be
faster on dense meshes but triangle based code gives tighter bounds on
scenes with non-uniform triangle sizes so we use it for now.