clusterizer: Allow larger leaves in kdtree
Instead of restricting leaves to just have a single point we now support
an arbitrary number of points. This can noticeably improve the overall
performance, e.g. on Bistro this change doubles the performance of
meshlet construction with leaf size 8.

Larger leaf sizes are also practical but for now to be conservative
we'll just set this to 8.