clusterizer: Make kdtree construction more robust
It's incorrect to adjust the median arbitrarily in general, as this may
not conform to the split value.

This change replaces the old logic by a new variant that's more robust
in general - when the partitioning returns results that are overly
biased towards either end, as determined by leaf_size, we return a
single leaf with all the points.

Note that this has an effect of making the average leaf size a bit larger
since the partitioning of nodes between 1.0 and 1.5x leaf_size can be
classified as uneven with a reasonably high probability.