clusterizer: Switch to variance-based split axis
Instead of using the bounding box size, which is a bit more prone to be
affected by one-off outliers, use variance to pick the axis; Welford's
online algorithm is used to compute both mean & variance in one pass.

The precision of computing mean is also higher compared to our old naive
computation as it doesn't suffer from catastrophic cancellation as much.